# api-doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "budgeting-backend",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/budgets": {
      "get": {
        "tags": [
          "Budgets"
        ],
        "operationId": "getBudgets",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Budget"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Budgets"
        ],
        "operationId": "createBudget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/budgets/{budget_id}": {
      "put": {
        "tags": [
          "Budgets"
        ],
        "operationId": "updateBudget",
        "parameters": [
          {
            "name": "budget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Budget": {
        "type": "object",
        "required": [
          "id",
          "name",
          "user_id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "target": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BudgetTarget"
              }
            ],
            "nullable": false
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "BudgetTarget": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "target_amount",
              "type"
            ],
            "properties": {
              "target_amount": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "OneTime"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "target_amount",
              "repeating_type",
              "schedule",
              "type"
            ],
            "properties": {
              "repeating_type": {
                "$ref": "#/components/schemas/RepeatingTargetType"
              },
              "schedule": {
                "$ref": "#/components/schemas/Schedule"
              },
              "target_amount": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Repeating"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "CreateBudgetRequest": {
        "type": "object",
        "required": [
          "name",
          "user_id"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "target": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateBudgetTargetRequest"
              }
            ],
            "nullable": false
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateBudgetTargetRequest": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "target_amount",
              "type"
            ],
            "properties": {
              "target_amount": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "OneTime"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "target_amount",
              "repeating_type",
              "schedule",
              "type"
            ],
            "properties": {
              "repeating_type": {
                "$ref": "#/components/schemas/RepeatingTargetType"
              },
              "schedule": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              },
              "target_amount": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Repeating"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "CreateScheduleRequest": {
        "type": "object",
        "required": [
          "period"
        ],
        "properties": {
          "period": {
            "$ref": "#/components/schemas/SchedulePeriod"
          }
        }
      },
...
      "RepeatingTargetType": {
        "type": "string",
        "enum": [
          "BuildUpTo",
          "RequireRepeating"
        ]
      },
      "Schedule": {
        "type": "object",
        "required": [
          "id",
          "period"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "period": {
            "$ref": "#/components/schemas/SchedulePeriod"
          }
        }
      },
      "SchedulePeriod": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "starting_on",
              "type"
            ],
            "properties": {
              "starting_on": {
                "type": "string",
                "format": "date"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Weekly"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "starting_on",
              "type"
            ],
            "properties": {
              "starting_on": {
                "type": "string",
                "format": "date"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Fortnightly"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "starting_on",
              "type"
            ],
            "properties": {
              "starting_on": {
                "type": "string",
                "format": "date"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Monthly"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "starting_on",
              "type"
            ],
            "properties": {
              "starting_on": {
                "type": "string",
                "format": "date"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Yearly"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "period",
              "every_x_periods",
              "type"
            ],
            "properties": {
              "every_x_periods": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              },
              "period": {
                "$ref": "#/components/schemas/SchedulePeriodType"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Custom"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
      "SchedulePeriodType": {
        "type": "string",
        "enum": [
          "Weekly",
          "Fortnightly",
          "Monthly",
          "Yearly"
        ]
      },
...
      "UpdateBudgetRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "target": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UpdateBudgetTargetRequest"
              }
            ],
            "nullable": false
          }
        }
      },
      "UpdateBudgetTargetRequest": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "target_amount",
              "type"
            ],
            "properties": {
              "target_amount": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "OneTime"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "target_amount",
              "repeating_type",
              "schedule",
              "type"
            ],
            "properties": {
              "repeating_type": {
                "$ref": "#/components/schemas/RepeatingTargetType"
              },
              "schedule": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              },
              "target_amount": {
                "type": "number",
                "format": "float"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Repeating"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "UpdateScheduleRequest": {
        "type": "object",
        "required": [
          "period"
        ],
        "properties": {
          "period": {
            "$ref": "#/components/schemas/SchedulePeriod"
          }
        }
      },
...
    }
  }
}